Predict Function Values at a Test Point

Predict Function Values at a Test Point

Usage

# S3 method for class 'gaussian_process'
predict(object, newdata = NULL, kernel_str = NA, ...)

Arguments

object

An object for which predictions are desired.

newdata

a new data set to generate predictions for. Must have all of the predictor columns of the original data set or an error will be returned.

kernel_str

the label of the kernel you would like to extract predictions for. This can be useful if you are (for example) attempting to isolate a specific signal from a data set such as a periodic component from a time series. Use the print method of the Gaussian process to determine the labels for the individual kernels.

...

additional arguments (ignored)